updates to lesson 1 snippets
