+++ tutorials/nservicebus-step-by-step/1-getting-started/Snippets/Core_9/Core_9.csproj
